셰이더에 유니폼 변수로 포인트 크기를 설정하는 기능 추가

+++ GameEngine/GLSL/Geometry2D.vert
@@ -6,11 +6,12 @@ layout(location = 1) in vec4 inColor;
 layout(location = 0) out vec4 outColor;
 
 uniform mat4 ortho;
-uniform int pointSize;
+uniform float pointSize;
 
 void main()
 {
 	gl_Position = ortho * vec4(inPosition, 1.0f);
 	gl_PointSize = pointSize;
+
 	outColor = inColor;
 }
